feat: Add dynamic transformation and logging to training script
- Integrated dynamic transformations with 5% increments from 5% to 80% for each image during training.
- Implemented the ability to specify composed transformations via CLI using the --composed_transforms argument.
- Enhanced MultiDirectoryDataSequence class to handle both base and composed transformations dynamically.
- Modified training script (train_dave2.py) to parse composed transformations from CLI and pass them to the dataset class.
- Added logging to track the number of images processed each epoch and significant training events.
- Included logging setup in train_dave2.py to output logs to training_log.log.
- Counted and logged the number of images processed during each epoch for verification.
- Updated __getitem__ method in MultiDirectoryDataSequence to apply transformations and composed transformations.
- Ensured all changes are backward compatible and maintain existing functionality.

Test:
- Verified the application of transformations and composed transformations.
- Ensured logs are correctly written and reflect the training process.
