+++ training/requirements.txt
@@ -1,7 +1,7 @@
 numpy~=1.19.5
 torch~=1.9.0+cu111
 opencv-python~=4.5.1.48
-torchvision~=0.10.0+cu111
+torchvision~=0.10.0
 scipy~=1.4.1
 matplotlib~=3.3.3
 tensorflow~=2.4.1
